fix: settings icon lives on each tab's own top bar, not a duplicate wrapping bar (tc-4jjw)
